Refactor: clean up zsh configuration files
Bug fixes:
- Fix typo: git dif → git diff
- Remove dangerous sz alias (source ~/.zshrc)
- Remove obsolete brew cask alias

Code cleanup:
- Delete commented proxy functions (now using zsh-proxy plugin)
- Delete prezto references and init directory
- Remove auto-push git functions (gwa, gwf, gwr, gwi, gwu, gwe)
- Remove install_fzf function (use brew instead)
- Remove fzf startup check (annoying message)

Configuration optimization:
- Remove duplicate PATH entries
- Remove obsolete paths (mysql@5.6, openresty, go11)
- Move LC_ALL from golang.zsh to env.zsh
- Rename config.toml → starship.toml (clearer naming)
- Reorganize macOS-specific aliases

Result: ~200 lines removed, 2.5KB saved, improved maintainability

+++ zsh/.config/zsh/alias.zsh
@@ -1,67 +1,36 @@
-
 # docker
 alias d='docker'
 alias dc='docker-compose'
 alias dcup='docker-compose up'
 alias da='docker attach'
-alias lima='limactl'
 
 # git
-alias gd='git dif'
+alias gd='git diff'
 alias ga='git add .'
 alias gc='git commit -m'
 alias g='gitui'
 alias gbl='git show-branch --list'
 alias gst='git status'
-alias rm='rm -i'
-#alias chmod='chmod --preserve-root'
 alias gfo='git fetch origin'
 
-
-# proxy - DEPRECATED: Now using zsh-proxy plugin
-# See: zmodule SukkaW/zsh-proxy in ~/.zimrc
-# Configuration in ~/.zshrc: PROXY_HTTP and PROXY_SOCKS5
-# Usage: proxy / noproxy
-#
-# proxy () {
-#   export https_proxy=http://127.0.0.1:6152
-#   export https_proxy=http://127.0.0.1:6152
-#   export all_proxy=socks5://127.0.0.1:6153
-#   echo "Proxy on"
-# }
-# noproxy () {
-#   unset http_proxy
-#   unset https_proxy
-#   unset all_proxy
-#   echo "Proxy off"
-# }
-
-
-# Common Aliases
-#
-# Overwrite Aliases
-# see also: https://github.com/sorin-ionescu/prezto/blob/master/modules/utility/init.zsh
+# common
+alias rm='rm -i'
 alias l='ls -alh'
 alias lm='ls -alh | "$PAGER"'
+alias ll='ls -lG'
+alias now='date +%s'
+alias j='z'
 
-# alias bat to cat
-# https://github.com/sharkdp/bat
+# bat (better cat)
 if command -v bat &> /dev/null; then
     alias cat='bat --theme OneHalfDark --style=plain --paging=never'
 fi
 
-
-# others
-alias now='date +%s'
-alias sz="source $HOME/.zshrc"
-alias j='z'
-alias ll='ls -lG'
-# alias rg='rg --column --line-number --hidden --sort path --no-heading --color=always --smart-case -- '
-
-# macos only
-alias refresh-dns='sudo killall -HUP mDNSResponder'
-
-# wechat 
-alias wechat='nohup /Applications/WeChat.app/Contents/MacOS/WeChat > /dev/null 2>&1 &'
 # lazygit
 alias lg="lazygit"
+
+# macOS specific
+if [[ "$(uname)" == "Darwin" ]]; then
+    alias refresh-dns='sudo killall -HUP mDNSResponder'
+    alias wechat='nohup /Applications/WeChat.app/Contents/MacOS/WeChat > /dev/null 2>&1 &'
+fi
